Rename files and update composer.json.
Shell files moved around, update upgrade tool to reflect those changes.
