+++ bin/cake.php
@@ -16,6 +16,6 @@
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-require dirname(dirname(__DIR__)) . '/config/bootstrap.php';
+require dirname(__DIR__) . '/config/bootstrap.php';
 
 exit(Cake\Console\ShellDispatcher::run($argv));
