+++ src/Shell/UpgradeShell.php
@@ -12,7 +12,7 @@
  * @since         CakePHP(tm) v 2.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Upgrade\Console\Command;
+namespace Cake\Upgrade\Shell;
 
 use Cake\Console\Shell;
 use Cake\Core\App;
