+++ tests/TestCase/Shell/UpgradeShellTest.php
